Support futures_codec as well as tokio-codec
This adds the possibility to use this codec in order to frame AsyncRead/AsyncWrite from the
futures 0.3 library as well as tokio's version.

Actual implementation of this is a single line. It just requires importing the `Encoder` and `Decoder`
traits from futures_code instead of from tokio-codec. I created an optional dependency
on futures_codec for this.

I have added documentation and and example.

A note on the dev-dependency on futures-preview. I have pinned it to version 0.3.0-alpha.17
because the runtime crate released on crates.io will not compile with futures 0.3.0-alpha.18,
however that last one was released yesterday, so I think this will be fixed soon after which
the pin can be removed.